% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{freestyler}
\alias{freestyler}
\title{Function to create Word report paragraph element}
\usage{
freestyler(bodytext = "Lorem Ipsium", text.align = "left")
}
\arguments{
\item{bodytext}{text of paragraph}

\item{text.align}{alignment of text}
}
\value{
Word report paragraph element
}
\description{
Function to create Word report paragraph element
}
