% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{genreport}
\alias{genreport}
\title{Function to create Word report}
\usage{
genreport(
  rep_cont = NULL,
  template = file.path(".", "templates", "word", "report_template_DECLS_v2.docx"),
  title = "Data Preparation Report",
  creator = "Michael Wild",
  target = "./DataPrepReportYemen.docx"
)
}
\arguments{
\item{rep_cont}{list of report content, see details for the structure of the required list}

\item{template}{path to template}

\item{title}{title of report}

\item{creator}{creator of report}

\item{target}{path to target}
}
\value{
Word report in target path
}
\description{
Function to create Word report
}
\details{
The list rep_cont must contain the following elements:
\itemize{
 \item doc_title: title of report
 \item sec_title: list of section titles
 \item sec_para: list of section paragraphs
 \item sec_table: list of section tables
 \item sec_graph: list of section graphs
 }
 sec_para, sec_table and sec_graph subsequently require a list of paragraphs, tables and graphs respectively,
 named para1, para2, para3, etc.
}
