% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{report_title}
\alias{report_title}
\title{Function to create Word report title element}
\usage{
report_title(qTitle)
}
\arguments{
\item{qTitle}{title of report}
}
\value{
Word report title element
}
\description{
Function to create Word report title element
}
